test(accordion): recategorize and clean up browser tests
Recategorize, prune, and consolidate the browser tests in
packages/react/src/components/accordion according to the rule from #7139.

Closes #7156
