--- packages/react/src/components/accordion/use-accordion.test.ts
@@ -1,119 +1,0 @@
-import type { PropsWithChildren, Ref } from "react"
-import { createElement, createRef } from "react"
-import { renderHook } from "#test"
-import {
-  AccordionContext,
-  AccordionDescendantsContext,
-  useAccordion,
-  useAccordionItem,
-} from "./use-accordion"
-
-function invokeCallbackRef<T>(ref: Ref<T> | undefined, node: null | T) {
-  if (typeof ref === "function") ref(node)
-}
-
-function invokeHandler<E>(handler: ((event: E) => void) | undefined, event: E) {
-  handler?.(event)
-}
-
-function AccordionItemWrapper({ children }: PropsWithChildren) {
-  const { descendants, getRootProps: _, ...context } = useAccordion()
-
-  return createElement(
-    AccordionDescendantsContext.Provider,
-    { value: descendants },
-    createElement(AccordionContext.Provider, { value: context }, children),
-  )
-}
-
-describe("useAccordion getRootProps", () => {
-  test("merges hook rest with user props via mergeProps", () => {
-    const hookRef = createRef<HTMLDivElement>()
-    const userRef = createRef<HTMLDivElement>()
-    const hookOnClick = vi.fn()
-    const userOnClick = vi.fn()
-    const { result } = renderHook(
-      () =>
-        useAccordion({
-          id: "hook-id",
-          ref: hookRef,
-          className: "hook",
-          style: { color: "red" },
-          "data-testid": "acc-root",
-          onClick: hookOnClick,
-        }),
-      { withProvider: false },
-    )
-
-    const merged = result.current.getRootProps({
-      id: "user-id",
-      ref: userRef,
-      className: "user",
-      style: { backgroundColor: "blue" },
-      onClick: userOnClick,
-    })
-    const node = document.createElement("div")
-    const event = new MouseEvent("click")
-
-    expect(typeof merged.ref).toBe("function")
-    invokeCallbackRef(merged.ref, node)
-    invokeHandler(merged.onClick, event as never)
-
-    expect(merged.id).toBe("user-id")
-    expect(String(merged.className)).toContain("hook")
-    expect(String(merged.className)).toContain("user")
-    expect(merged.style).toStrictEqual({
-      backgroundColor: "blue",
-      color: "red",
-    })
-    expect(hookRef.current).toBe(node)
-    expect(userRef.current).toBe(node)
-    expect(hookOnClick).toHaveBeenCalledWith(event)
-    expect(userOnClick).toHaveBeenCalledWith(event)
-    expect(merged["data-testid"]).toBe("acc-root")
-  })
-})
-
-describe("useAccordionItem getItemProps", () => {
-  test("merges hook rest and user props via mergeProps", () => {
-    const hookRef = createRef<HTMLDivElement>()
-    const userRef = createRef<HTMLDivElement>()
-    const hookOnClick = vi.fn()
-    const userOnClick = vi.fn()
-    const { result } = renderHook(
-      () =>
-        useAccordionItem({
-          ref: hookRef,
-          className: "hook",
-          style: { color: "red" },
-          index: 0,
-          onClick: hookOnClick,
-        }),
-      { withProvider: false, wrapper: AccordionItemWrapper },
-    )
-
-    const merged = result.current.getItemProps({
-      ref: userRef,
-      className: "user",
-      style: { backgroundColor: "blue" },
-      onClick: userOnClick,
-    })
-    const node = document.createElement("div")
-    const event = new MouseEvent("click")
-
-    expect(typeof merged.ref).toBe("function")
-    invokeCallbackRef(merged.ref, node)
-    invokeHandler(merged.onClick, event as never)
-
-    expect(String(merged.className)).toContain("hook")
-    expect(String(merged.className)).toContain("user")
-    expect(merged.style).toStrictEqual({
-      backgroundColor: "blue",
-      color: "red",
-    })
-    expect(hookRef.current).toBe(node)
-    expect(userRef.current).toBe(node)
-    expect(hookOnClick).toHaveBeenCalledWith(event)
-    expect(userOnClick).toHaveBeenCalledWith(event)
-  })
-})
